feat(auth): enterprise auth provider changes
